initial typescript implementation of data types
- some areas are still 'any', and need to be fixed
- npm audit fix
