+++ app.json
@@ -7,7 +7,7 @@
     "platforms": [
       "android"
     ],
-    "version": "1.1.6",
+    "version": "1.1.7",
     "orientation": "portrait",
     "icon": "./assets/images/icon.png",
     "splash": {
@@ -21,7 +21,7 @@
     "android": {
       "package": "com.kayakfreestyle.kayakfreestyleapp",
       "permissions": [],
-      "versionCode": 7
+      "versionCode": 9
     },
     "assetBundlePatterns": [
       "**/*"
