+++ data/moves_lists/move_list.json
@@ -193,7 +193,7 @@
             "Huge": 50,
             "Link": 30,
             "Reverse": true
-        },
+        }
     ],
     "wave": [
         {
