+++ screens/mycomponents/entryDynamicButton.js
@@ -29,10 +29,7 @@ const DynamicButtonPresentation = React.memo((props) => {
             props.direction
         ].scored == false
     ) {
-        const buttonName =
-            oneSidedMoves.indexOf(props.move.Move) > -1
-                ? props.move.Move
-                : props.move.Move + " " + props.direction;
+        const buttonName = props.move.Move
         return (
             <Button
                 onPress={_handleMove(
@@ -74,7 +71,7 @@ const DynamicButtonPresentation = React.memo((props) => {
                         buttonStyle={thisMove[props.direction]["superClean"] ? styles.bonusScored : styles.noBonus}
                     />
                 </View>
-                <View style={{ width: "25%" }}>
+                <View style={{ width: "33%" }}>
                     <Button
                         onPress={_handleMove(props.paddler, props.run, props.move.Move, props.direction, "air")}
                         title={"A"}
@@ -82,7 +79,7 @@ const DynamicButtonPresentation = React.memo((props) => {
                         buttonStyle={thisMove[props.direction]["air"] ? styles.bonusScored : styles.noBonus}
                     />
                 </View>
-                <View style={{ width: "40%" }}>
+                <View style={{ width: "33%" }}>
                     <Button
                         onPress={_handleMove(props.paddler, props.run, props.move.Move, props.direction, "huge")}
                         title={"H"}
@@ -90,7 +87,7 @@ const DynamicButtonPresentation = React.memo((props) => {
                         buttonStyle={thisMove[props.direction]["huge"] ? styles.bonusScored : styles.noBonus}
                     />
                 </View>
-                <View style={{ width: "35%" }}>
+                <View style={{ width: "33%" }}>
                     <Button
                         onPress={_handleMove(props.paddler, props.run, props.move.Move, props.direction, "link")}
                         title={"L"}
